Fix s3 bucket creation and account-ready polling for reuse test, fix account clean up for both account creation and reuse tests

+++ test/integration/tests/test_nonccs_account_reuse.sh
@@ -5,11 +5,15 @@ source test/integration/test_envs
 EXIT_TEST_FAIL_ACCOUNT_PROVISIONING_FAILED=2
 EXIT_TEST_FAIL_REUSED_ACCOUNT_NOT_READY=3
 EXIT_TEST_FAIL_ACCOUNT_NOT_REUSED=4
+EXIT_TEST_FAIL_SECRET_INVALID_CREDS=5
+EXIT_TEST_FAIL_S3_BUCKET_CREATION=6
 
 declare -A exitCodeMessages
 exitCodeMessages[$EXIT_TEST_FAIL_ACCOUNT_PROVISIONING_FAILED]="Test Account CR has a status of failed. Check AAO logs for more details."
 exitCodeMessages[$EXIT_TEST_FAIL_REUSED_ACCOUNT_NOT_READY]="Test Account CR is not in a ready state. Check AAO logs for more details."
 exitCodeMessages[$EXIT_TEST_FAIL_ACCOUNT_NOT_REUSED]="Test Account CR was not reused. Check AAO logs for more details."
+exitCodeMessages[$EXIT_TEST_FAIL_SECRET_INVALID_CREDS]="Test Account secret credentials are invalid. Check AAO logs for more details."
+exitCodeMessages[$EXIT_TEST_FAIL_S3_BUCKET_CREATION]="Failed to create AWS S3 bucket. Check logs for more details."
 
 function setupTestPhase {
   # move OSD Staging 1 account to root ou to avoid ChildNotFoundInOU errors
@@ -17,57 +21,50 @@ function setupTestPhase {
 
   oc process --local -p NAME="${ACCOUNT_CLAIM_NAMESPACE}" -f hack/templates/namespace.tmpl | oc apply -f -
 
-  oc get account "${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}" -n "${NAMESPACE}" 2>/dev/null
-  ACCOUNT_CR_EXISTS=$?
-
-  if [ $ACCOUNT_CR_EXISTS -ne 0 ]; then
+  if ! oc get account "${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}" -n "${NAMESPACE}" 2>/dev/null; then
       echo "Creating Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}"
       if ! oc process -p AWS_ACCOUNT_ID="${OSD_STAGING_1_AWS_ACCOUNT_ID}" -p ACCOUNT_CR_NAME="${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}" -p NAMESPACE="${NAMESPACE}" -f hack/templates/aws.managed.openshift.io_v1alpha1_account.tmpl | oc apply -f -; then
           echo "Failed to create account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}"
           exit "$EXIT_FAIL_UNEXPECTED_ERROR"
       fi
-  else
-
-      if ! STATUS=$(oc get account "${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}" -n "${NAMESPACE}" -o json | jq -r '.status.state'); then
-          echo "Failed to get status of account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}"
-          exit "$EXIT_FAIL_UNEXPECTED_ERROR"
-      fi
-
-      if [ "$STATUS" == "Ready" ]; then
-          echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} is ready."
-      elif [ "$STATUS" == "Failed" ]; then
-          echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} failed to create"
-          exit $EXIT_TEST_FAIL_ACCOUNT_PROVISIONING_FAILED
-      else
-          echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} status is ${STATUS}, waiting for it to become ready or fail."
-          exit "$EXIT_RETRY"
-      fi
   fi
 
-  oc get accountclaim "${ACCOUNT_CLAIM_NAME}" -n "${ACCOUNT_CLAIM_NAMESPACE}" 2>/dev/null
-  ACCOUNT_CLAIM_CR_EXISTS=$?
+  if ! STATUS=$(oc get account "${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}" -n "${NAMESPACE}" -o json | jq -r '.status.state'); then
+      echo "Failed to get status of account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}"
+      exit "$EXIT_FAIL_UNEXPECTED_ERROR"
+  fi
 
-  if [ $ACCOUNT_CLAIM_CR_EXISTS -ne 0 ]; then
+  if [ "$STATUS" == "Ready" ]; then
+      echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} is ready."
+  elif [ "$STATUS" == "Failed" ]; then
+      echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} failed to create"
+      exit $EXIT_TEST_FAIL_ACCOUNT_PROVISIONING_FAILED
+  else
+      echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} status is ${STATUS}, waiting for it to become ready or fail."
+      exit "$EXIT_RETRY"
+  fi
+
+  if ! oc get accountclaim "${ACCOUNT_CLAIM_NAME}" -n "${ACCOUNT_CLAIM_NAMESPACE}" 2>/dev/null; then
       echo "Creating Account Claim ${ACCOUNT_CLAIM_NAME}"
       if ! oc process --local -p NAME="${ACCOUNT_CLAIM_NAME}" -p NAMESPACE="${ACCOUNT_CLAIM_NAMESPACE}" -f hack/templates/aws.managed.openshift.io_v1alpha1_accountclaim_cr.tmpl | oc apply -f -; then
           echo "Failed to create account claim ${ACCOUNT_CLAIM_NAME}"
           exit "$EXIT_FAIL_UNEXPECTED_ERROR"
       fi
-  else
-      if ! STATUS=$(oc get accountclaim "${ACCOUNT_CLAIM_NAME}" -n "${ACCOUNT_CLAIM_NAMESPACE}" -o json | jq -r '.status.state'); then
-          echo "Failed to get status of account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}"
-          exit "$EXIT_FAIL_UNEXPECTED_ERROR"
-      fi
+  fi
 
-      if [ "$STATUS" == "Ready" ]; then
-          echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} is ready."
-      elif [ "$STATUS" == "Failed" ]; then
-          echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} failed to create"
-          exit $EXIT_TEST_FAIL_ACCOUNT_PROVISIONING_FAILED
-      else
-          echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} status is ${STATUS}, waiting for it to become ready or fail."
-          exit "$EXIT_RETRY"
-      fi
+  if ! STATUS=$(oc get accountclaim "${ACCOUNT_CLAIM_NAME}" -n "${ACCOUNT_CLAIM_NAMESPACE}" -o json | jq -r '.status.state'); then
+      echo "Failed to get status of account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}"
+      exit "$EXIT_FAIL_UNEXPECTED_ERROR"
+  fi
+
+  if [ "$STATUS" == "Ready" ]; then
+      echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} is ready."
+  elif [ "$STATUS" == "Failed" ]; then
+      echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} failed to create"
+      exit $EXIT_TEST_FAIL_ACCOUNT_PROVISIONING_FAILED
+  else
+      echo "Account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD} status is ${STATUS}, waiting for it to become ready or fail."
+      exit "$EXIT_RETRY"
   fi
 
   # Create S3 Bucket
@@ -77,9 +74,19 @@ function setupTestPhase {
   AWS_SECRET_ACCESS_KEY=$(oc get secret "${IAM_USER_SECRET}" -n "${NAMESPACE}" -o json | jq -r '.data.aws_secret_access_key' | base64 -d)
   export AWS_SECRET_ACCESS_KEY
 
-  aws s3api create-bucket --bucket "${REUSE_BUCKET_NAME}" --region=us-east-1
+  if [ -z "$AWS_ACCESS_KEY_ID" ] || [ -z "$AWS_SECRET_ACCESS_KEY" ]; then
+    echo "AWS credentials not found in secret"
+    exit $EXIT_TEST_FAIL_SECRET_INVALID_CREDS
+  fi
 
-  # TODO: Might need to remove accountclaim finalizers
+  REUSE_UUID="$(uuidgen | awk -F- '{ print tolower($$2) }')"
+  REUSE_BUCKET_NAME=test-reuse-bucket-${REUSE_UUID}
+
+  if ! aws s3api create-bucket --bucket "${REUSE_BUCKET_NAME}" --region=us-east-1; then
+      echo "Failed to create s3 bucket ${REUSE_BUCKET_NAME}."
+      exit $EXIT_TEST_FAIL_S3_BUCKET_CREATION
+  fi
+
   oc process --local -p NAME="${ACCOUNT_CLAIM_NAME}" -p NAMESPACE="${ACCOUNT_CLAIM_NAMESPACE}" -f hack/templates/aws.managed.openshift.io_v1alpha1_accountclaim_cr.tmpl | oc delete --now --ignore-not-found -f -
 
   # Delete reuse namespace
@@ -87,14 +94,18 @@ function setupTestPhase {
 }
 
 function cleanupTestPhase {
-    if ! oc get account "${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}" -n "${NAMESPACE}" 2>/dev/null; then
+    if oc get account "${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}" -n "${NAMESPACE}" 2>/dev/null; then
+        echo "Found account, removing it now."
         oc patch account "${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}" -n "${NAMESPACE}" -p '{"metadata":{"finalizers":null}}' --type=merge
         oc process -p AWS_ACCOUNT_ID="${OSD_STAGING_1_AWS_ACCOUNT_ID}" -p ACCOUNT_CR_NAME="${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}" -p NAMESPACE="${NAMESPACE}" -f hack/templates/aws.managed.openshift.io_v1alpha1_account.tmpl | oc delete --now --ignore-not-found -f -
+        echo "Should have deleted account?"
 
         if ! oc get account "${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}" -n "${NAMESPACE}" 2>/dev/null; then
             echo "Failed to delete account ${OSD_STAGING_1_ACCOUNT_CR_NAME_OSD}"
             exit "$EXIT_FAIL_UNEXPECTED_ERROR"
         fi
+
+        echo "Clean up definitely should have worked by here."
     fi
 
     exit "$EXIT_PASS"
